+++ dump.h
@@ -1,4 +1,4 @@
 void dump_begin(int dump);
 void dump_end(int dump);
-void dump_out(int dump, unsigned int *x, unsigned int *y, int components, int metabits, int meta);
+void dump_out(int dump, unsigned int *x, unsigned int *y, int components, int metabits, long long meta);
